// src/api/generated/models/PaymentMethodTableDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Format for Payment Method read
 * @export
 * @interface PaymentMethodTableDTO
 */
export interface PaymentMethodTableDTO {
    /**
     * 
     * @type {string}
     * @memberof PaymentMethodTableDTO
     */
    createdById?: string;
    /**
     * 
     * @type {string}
     * @memberof PaymentMethodTableDTO
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof PaymentMethodTableDTO
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof PaymentMethodTableDTO
     */
    updatedById?: string;
}

/**
 * Check if a given object implements the PaymentMethodTableDTO interface.
 */
export function instanceOfPaymentMethodTableDTO(value: object): value is PaymentMethodTableDTO {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function PaymentMethodTableDTOFromJSON(json: any): PaymentMethodTableDTO {
    return PaymentMethodTableDTOFromJSONTyped(json, false);
}

export function PaymentMethodTableDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): PaymentMethodTableDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'createdById': json['createdById'] == null ? undefined : json['createdById'],
        'id': json['id'],
        'name': json['name'],
        'updatedById': json['updatedById'] == null ? undefined : json['updatedById'],
    };
}

export function PaymentMethodTableDTOToJSON(json: any): PaymentMethodTableDTO {
    return PaymentMethodTableDTOToJSONTyped(json, false);
}

export function PaymentMethodTableDTOToJSONTyped(value?: PaymentMethodTableDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'createdById': value['createdById'],
        'id': value['id'],
        'name': value['name'],
        'updatedById': value['updatedById'],
    };
}